demote @Layout to sample
